rewrite-part1
