+++ .gitignore
@@ -12,6 +12,4 @@
 AbyssDiver.html
 *.pyc
 *.exe
-log.txt
-device
-/local-gen/models
+*.log
