+++ build-linux-macos.sh
@@ -4,38 +4,34 @@
 set -e
 
 # Function to check for Python installation
-check_and_install_python() {
+check_python() {
 	echo "Checking for Python installation..."
 	if ! command -v python &>/dev/null && ! command -v python3 &>/dev/null && ! command -v py &>/dev/null; then
-		echo "Python is not installed. Attempting to install Python..."
-		chmod +x install_python_linux_macos.sh
-		./install_python_linux_macos.sh
-	else
-		echo "Python is already installed."
+		echo "Python is not installed. Please install manually and restart the terminal."
+		echo "Press enter to exit."
+		read -p ""
+		exit 1
 	fi
 }
 
 # Function to ensure pip is installed and up-to-date
 install_or_update_pip() {
 	echo "Ensuring pip is installed and up-to-date..."
-	if command -v python &>/dev/null; then
+	if command -v py &>/dev/null; then
+		py -m ensurepip --upgrade || true
+		py -m pip install --upgrade pip
+	elif command -v python &>/dev/null; then
 		python -m ensurepip --upgrade || true
 		python -m pip install --upgrade pip
 	elif command -v python3 &>/dev/null; then
 		python3 -m ensurepip --upgrade || true
 		python3 -m pip install --upgrade pip
-	elif command -v py &>/dev/null; then
-		py -m ensurepip --upgrade || true
-		py -m pip install --upgrade pip
-	else
-		echo "Error: Unable to locate Python or pip."
-		exit 1
 	fi
 }
 
 # Main function for execution
 main() {
-	check_and_install_python
+	check_python
 	install_or_update_pip
 
 	# Run the Python build script with `-w` argument
