--- local-gen/requirements.txt
@@ -1,9 +1,0 @@
-# packages required by the proxy (auto-installed)
-aiohttp
-fastapi
-patool
-pillow
-pydantic
-requests_toolbelt
-uvicorn[standard]
-websocket-client
