+++ pyrightconfig.json
@@ -1,0 +1,23 @@
+{
+	"include": [
+		"src"
+	],
+	"exclude": [
+		"comfyui/tools",
+		"local-gen/tools",
+		"images/"
+	],
+	"ignore": [
+		"comfyui/tools",
+		"local-gen/tools",
+		"images/"
+	],
+	"defineConstant": {
+		"DEBUG": true
+	},
+	"reportMissingImports": "error",
+	"reportMissingTypeStubs": false,
+	"pythonVersion": "3.10",
+	"pythonPlatform": "Windows",
+	"executionEnvironments": []
+}
